requirements.txt
# File: requirements.txt
langgraph>=0.2
google-generativeai>=0.8
requests>=2.31
rich>=13.0
python-dotenv>=1.0
